moving AgentSelection to its own component
